Restore old `nl` translation as `nl-BE`
As result of the discussion in #168 this restores the "old" Dutch
translation as `nl-BE` as it seems there are major differences between
the Dutch spoken in NL and the one in BE. This way the differences in
the language are used for the different countries.

The `nl-NL` translation is used as "standard NL" as there are plenty
more `nl-*` language keys. Those are served by the translation for the
Netherlands.

Signed-off-by: Knut Ahlers <knut@ahlers.me>