feat: add expiryChoicesHuman, remove 'default expiry' from dropdown
Focussing on front-end UI, 'default expiry' means nothing. Rather than relying on non-human readable, unix timestamps, introducing human readable options: (s)econds, (m)inutes, (h)ours, (d)ays. The defaultExpiryHuman should be any of the options for expiryChoicesHuman. If it's not, it'll fallback to `null`, which will fallback to the `defaultExpiry` as it is now

Save the selected option as a cookie, only after initially loaded. Load from cookie on page load. API is not changed, still accepts seconds